correction de certaines erreurs

+++ include/tools.h
@@ -12,7 +12,7 @@ xt::xarray<float> normalized(xt::xarray<float> input);
 
 xt::xarray<float> flatten(xt::xarray<float> input);
 
-float MSE(xt::xarray<float> output, xt::xarray<int> trueValue);
+float MSE(xt::xarray<float> output, xt::xarray<float> trueValue);
 
 float crossEntropy(xt::xarray<float> output, xt::xarray<int> trueValue);
